Initial Commit
Sorry for the inactivity been busy with holidays, I have reworked
priorities with the monstercat suite and have decided to work on getting
the music plugin up and running. I have blocked out code to duplicate
original NowPlaying plugin as well as the Web plugin I intend to see if
I can make. The web plugin may be scrapped if not possible.