Fixed a bug with lastInfo array
I had set the latestInfo array to static, while it may be useful to set
this to static in the future and just have a parent update it it is not
useful right now. Now it is now set to static so each source of info
will correctly display now.
Also I changed the flow of the dynamic music reader to prevent a cycle
from happening by having it always run all the way through, this is
great on the first run but makes for extra work on further runs. Ill
look into a better implementation in the future.

I have built out a simple executable check to see if any source should
be displayed, still in the testing.