Attempts to get websocket running
Trying to use off the shelf stuff for websocket and JSON handling but
whenever I try to use functions in a way that wont be stripped by the
comipler (I thought I ahd it working for a bit but realized it was a
line that did nothing and likely the comiple optimizer stipped it out) I
have a feeling I am just missing something dumb and the namespaces are
not being found at run time. I may just look at this on Saturday (I have
a busy Friday) with a fresh cup of coffee.