GPM title functionallity test.
Now correctly opens a websocket and handles the title data from the JSON
response of GPMDP. Need to clean it up and close and dealoc the
websocket but it works. Gonna get all the info from GPMDP ported in to
the correct info types and then I will work on the controls.