More performance improvements
Now unsupported types are not serialized and ignored. It seems like some
measures are sharing a socket and some are not, have a feeling it has to
do with threading. Going to look into seeing what I can do to have them
all share one thread (Likely will just have to have a parent and child
thread)