Improved Performance
Performance is now at acceptable levels. Turns out that attempting to
open a bad websocket is less overhead than getProcessByName which makes
sense. I may migrate to not allowing it to try to open a websocket more
than every 10 seconds maybe in the future, however right now it is nice
as it just does it at whatever the update rate of the measure is. Maybe
I will make it a optional flag.