Fixed a bug with default cover location
I forgot to set the cover location to default while the image is
downloading, this prevented OnChange's from firing at all. I have set it
before the thread is started to maxamize time between when the default
cover is set and when the actual will be so that OnChange is likely to
fire.