PlayPause Next Previous support
The support is a little awkward at the moment as I am fairly certain
that the authcode will not work across machines. Gonna see if I can find
someone to test that, if it does not I will have to see if I can get the
inital code out of GPMDP using AutoIt or just allow the user to enter it
via a measure.