Authcode read write
Currently the authcode when read from the file never seems to be
accepted by GPMDP, at first I was thinking it had something to do with
the size being read or the state of the websocket when it sends but not
I just realized that the callback for the connection never happens, even
when I do get authorized. Investigating.

Also once this bug is fixed this code is gonna get cleaned up, then a
fork is gonna happen that will contain just GPMDP related stuff that
will be released as its own plugin.