Remote connection improvements
Any skin can now get the connection status. 0 is GPMDP is not
running/does not have network access. 1 is it is running but the remote
is not authenticated. 2 is it is connected and has remote access. I am
going to change the input skin in the future to change what it displays
based on this.

Also if the remote access code is not known at the time the plugin is
first used it sends a request to GPMDP. It does not however send it on
reload but since it will send when GPMDP is opened I dont think this is
an issue.