GPMDP Plguin Version 1.0
The constructor for the musicInfo object now initializes all variables.
ConnectionStatus now uses Status
Repeat and Shuffle support added for both reading current value and
setting it. Both toggle through all options, in the future I may add the
ability to set it.
Connection name for remote renamed from the testing string.
A check to make sure that the websocket needs to be closed and exists
has been added to the finalize to prevent rainmeter from crashing when a
skin is unloaded or exiting rainmeter when the websocket is not
connected.