Progress is now a double
Also I left in code where I was experimenting with getting genre. Turns
out GPMDP does not support it however I left it in so that I can reuse
it if I need to in the future which I will use some parts for volume
however that is going to function a bit differently.