Fixed memory leak when GPMDP is closed
A memory leak was intruduced when I switched from looking up if GPMDP is
opened before trying to reconnect to it. I have now done this check on a
different thread which is shared between all measures so only one thread
is opened. In the future I may rewrite the reconnect and reinit threads
to be more performant. (Thanks to herrerarausaure for finding this bug.

Also at the request of MarcoPixel I have changed position and duration
to initialize to 00:00 instead of blank.

The remaining commented out changes are tests of the queue system. I
need to work on a high performance low memory using solution to
searching through that string so I am just returning the previous and
next 10 songs.