Added support for reading custom colors on load
Since this is such an edge case I have it only being read on skin
refresh, if there is demand for this to change in the future I can make
to poll the file but I figure it is best to just read it on load. I am
also working on adding support for automatically authenticating myself
so an auth skin does not need to be included.