Added missing edge cases for settings file
It seems that GPMDP only adds settings to the .settings.json file that
have changed so it is possible I think for the authoerized_devices
seciton to be missing. I am unable to test this without doing a full
wipe of all GPMDP files so I added support for potential edge cases
anyways, it should now handle it no mater what.