Add AsDecimal and DisableLeadingZero support
Fixes issues #10 and #11
Slight change of flow when it comes to calculating position and
duration, they are now calculated during the update cycle instead of
during the websocket info send. This makes the duration string
calculated more often but should in theory make the position string
calculated less often. I am also tempted to make a 0 padding count
option in the future.