Start of new queue changes
New bang to change to a specific song in the GPM queue, currently is the
absolute position of the song in the queue will be relative later.
Broken in this build: Queue, whole thing literally hit the delete key on
it so dont use it. Will be back up and running in the next build