Reimplmented relative queue
Implments #9 and is as good as it is gonna get for #12. I would love to
just go to a location but it seems that is just not posible right now. I
still need to replicate the old null data when trying to go outside the
array instead of just giving the same data.